requirements.txt
gpiozero
rpi-gpio